add back predict.py
